requirements.txt
pyserial
numpy
neat